!!![TASK] Restructure Menu handling
