+++ packages/guides/src/Compiler/NodeTransformers/MenuNodeTransformers/TocNodeTransformer.php
@@ -1,0 +1,39 @@
+<?php
+
+declare(strict_types=1);
+
+namespace phpDocumentor\Guides\Compiler\NodeTransformers\MenuNodeTransformers;
+
+use phpDocumentor\Guides\Compiler\CompilerContext;
+use phpDocumentor\Guides\Compiler\NodeTransformer;
+use phpDocumentor\Guides\Nodes\Menu\TocNode;
+use phpDocumentor\Guides\Nodes\Node;
+
+use function assert;
+
+/** @implements NodeTransformer<TocNode> */
+final class TocNodeTransformer implements NodeTransformer
+{
+    public function enterNode(Node $node, CompilerContext $compilerContext): Node
+    {
+        assert($node instanceof TocNode);
+        $compilerContext->getDocumentNode()->addTocNode($node);
+
+        return $node;
+    }
+
+    public function leaveNode(Node $node, CompilerContext $compilerContext): Node|null
+    {
+        return $node;
+    }
+
+    public function supports(Node $node): bool
+    {
+        return $node instanceof TocNode;
+    }
+
+    public function getPriority(): int
+    {
+        return 1000;
+    }
+}
