+++ tests/Integration/tests/navigation/contents/contents-depth-1/expected/index.html
@@ -1,10 +1,3 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <title>Title</title>
-
-</head>
-<body>
 <!-- content start -->
     <div class="section" id="title">
             <h1>Title</h1>
@@ -12,6 +5,7 @@
             <div class="contents">
         <ul class="menu-level">
     <li class="toc-item"><a href="/index.html#depth-level-1">Depth Level 1</a></li>
+
     <li class="toc-item"><a href="/index.html#2-depth-level-1">2. Depth Level 1</a></li>
 
     </ul>
@@ -60,5 +54,3 @@
     </div>
 
 <!-- content end -->
-</body>
-</html>
