+++ tests/Integration/tests/navigation/contents/contents-depth-2/expected/index.html
@@ -1,17 +1,18 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <title>Title</title>
-
-</head>
-<body>
 <!-- content start -->
     <div class="section" id="title">
             <h1>Title</h1>
 
             <div class="contents">
         <ul class="menu-level">
-    <li class="toc-item"><a href="/index.html#depth-level-1">Depth Level 1</a></li>
+    <li class="toc-item"><a href="/index.html#depth-level-1">Depth Level 1</a>        <ul class="section-level-2">
+            <li class="toc-item"><a href="/index.html#depth-level-2">Depth Level 2</a></li>
+
+                    </ul></li>
+
+    <li class="toc-item"><a href="/index.html#2-depth-level-1">2. Depth Level 1</a>        <ul class="section-level-2">
+            <li class="toc-item"><a href="/index.html#2-depth-level-2">2. Depth Level 2</a></li>
+
+                    </ul></li>
 
     </ul>
 </div>
@@ -59,5 +60,3 @@
     </div>
 
 <!-- content end -->
-</body>
-</html>
