+++ tests/Integration/tests/navigation/contents/contents-depth-3/expected/index.html
@@ -1,28 +1,68 @@
 <!-- content start -->
-<div class="section" id="title">
-    <h1>Title</h1>
+    <div class="section" id="title">
+            <h1>Title</h1>
 
-    <div class="contents"><ul class="menu-level"><li class="toc-item"><a href="/contents-depth-3.html#depth-level-1">Depth Level 1</a><ul class="section-level-1"><li class="toc-item"><a href="/contents-depth-3.html#depth-level-2">Depth Level 2</a><ul class="section-level-1"><li class="toc-item"><a href="/contents-depth-3.html#depth-level-3">Depth Level 3</a></li></ul></li></ul></li></ul></div>
-    <div class="section" id="depth-level-1">
-        <h2>Depth Level 1</h2>
+            <div class="contents">
+        <ul class="menu-level">
+    <li class="toc-item"><a href="/index.html#depth-level-1">Depth Level 1</a>        <ul class="section-level-2">
+            <li class="toc-item"><a href="/index.html#depth-level-2">Depth Level 2</a>        <ul class="section-level-2">
+            <li class="toc-item"><a href="/index.html#depth-level-3">Depth Level 3</a></li>
 
-        <div class="section" id="depth-level-2">
+                    </ul></li>
+
+                    </ul></li>
+
+    <li class="toc-item"><a href="/index.html#2-depth-level-1">2. Depth Level 1</a>        <ul class="section-level-2">
+            <li class="toc-item"><a href="/index.html#2-depth-level-2">2. Depth Level 2</a>        <ul class="section-level-2">
+            <li class="toc-item"><a href="/index.html#2-depth-level-3">2. Depth Level 3</a></li>
+
+                    </ul></li>
+
+                    </ul></li>
+
+    </ul>
+</div>
+
+            <div class="section" id="depth-level-1">
+            <h2>Depth Level 1</h2>
+
+            <div class="section" id="depth-level-2">
             <h3>Depth Level 2</h3>
 
             <div class="section" id="depth-level-3">
-                <h4>Depth Level 3</h4>
+            <h4>Depth Level 3</h4>
 
-                <div class="section" id="depth-level-4">
-                    <h5>Depth Level 4</h5>
-
-                </div>
-
-            </div>
-
-        </div>
+            <div class="section" id="depth-level-4">
+            <h5>Depth Level 4</h5>
 
     </div>
 
-</div>
+    </div>
+
+    </div>
+
+    </div>
+
+            <div class="section" id="2-depth-level-1">
+            <h2>2. Depth Level 1</h2>
+
+            <div class="section" id="2-depth-level-2">
+            <h3>2. Depth Level 2</h3>
+
+            <div class="section" id="2-depth-level-3">
+            <h4>2. Depth Level 3</h4>
+
+            <div class="section" id="2-depth-level-4">
+            <h5>2. Depth Level 4</h5>
+
+    </div>
+
+    </div>
+
+    </div>
+
+    </div>
+
+    </div>
 
 <!-- content end -->
