+++ tests/Integration/tests/navigation/contents/contents-depth-3/input/index.rst
@@ -4,6 +4,18 @@ Title
 .. contents::
     :depth: 3
 
+Depth Level 1
+-------------
+
+Depth Level 2
+~~~~~~~~~~~~~
+
+Depth Level 3
++++++++++++++
+
+Depth Level 4
+.............
+
 2. Depth Level 1
 ----------------
