+++ tests/Integration/tests/navigation/contents/contents-no-depth/expected/index.html
@@ -1,23 +1,24 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <title>Title</title>
-
-</head>
-<body>
 <!-- content start -->
     <div class="section" id="title">
             <h1>Title</h1>
 
             <div class="contents">
         <ul class="menu-level">
-    <li class="toc-item"><a href="/index.html#title">Title</a>        <ul class="section-level-1">
-            <li class="toc-item"><a href="/index.html#depth-level-1">Depth Level 1</a>        <ul class="section-level-2">
+    <li class="toc-item"><a href="/index.html#depth-level-1">Depth Level 1</a>        <ul class="section-level-2">
             <li class="toc-item"><a href="/index.html#depth-level-2">Depth Level 2</a>        <ul class="section-level-2">
             <li class="toc-item"><a href="/index.html#depth-level-3">Depth Level 3</a>        <ul class="section-level-3">
             <li class="toc-item"><a href="/index.html#depth-level-4">Depth Level 4</a></li>
 
                     </ul></li>
+
+                    </ul></li>
+
+                    </ul></li>
+
+    <li class="toc-item"><a href="/index.html#2-depth-level-1">2. Depth Level 1</a>        <ul class="section-level-2">
+            <li class="toc-item"><a href="/index.html#2-depth-level-2">2. Depth Level 2</a>        <ul class="section-level-2">
+            <li class="toc-item"><a href="/index.html#2-depth-level-3">2. Depth Level 3</a>        <ul class="section-level-3">
+            <li class="toc-item"><a href="/index.html#2-depth-level-4">2. Depth Level 4</a></li>
 
                     </ul></li>
 
@@ -48,8 +49,26 @@
 
     </div>
 
+            <div class="section" id="2-depth-level-1">
+            <h2>2. Depth Level 1</h2>
+
+            <div class="section" id="2-depth-level-2">
+            <h3>2. Depth Level 2</h3>
+
+            <div class="section" id="2-depth-level-3">
+            <h4>2. Depth Level 3</h4>
+
+            <div class="section" id="2-depth-level-4">
+            <h5>2. Depth Level 4</h5>
+
+    </div>
+
+    </div>
+
+    </div>
+
+    </div>
+
     </div>
 
 <!-- content end -->
-</body>
-</html>
