+++ tests/Integration/tests/navigation/contents/contents-no-depth/input/index.rst
@@ -14,3 +14,15 @@ Depth Level 3
 
 Depth Level 4
 .............
+
+2. Depth Level 1
+----------------
+
+2. Depth Level 2
+~~~~~~~~~~~~~~~~
+
+2. Depth Level 3
+++++++++++++++++
+
+2. Depth Level 4
+................
