+++ tests/Integration/tests/navigation/contents/contents-with-title/expected/index.html
@@ -1,10 +1,3 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <title>Title</title>
-
-</head>
-<body>
 <!-- content start -->
     <div class="section" id="title">
             <h1>Title</h1>
@@ -12,13 +5,10 @@
             <div class="contents">
         <p class="topic-title"><strong>Table of Content</strong></p>
         <ul class="menu-level">
-    <li class="toc-item"><a href="/index.html#title">Title</a>        <ul class="section-level-1">
-            <li class="toc-item"><a href="/index.html#depth-level-1">Depth Level 1</a>        <ul class="section-level-2">
+    <li class="toc-item"><a href="/index.html#depth-level-1">Depth Level 1</a>        <ul class="section-level-2">
             <li class="toc-item"><a href="/index.html#depth-level-2">Depth Level 2</a>        <ul class="section-level-2">
             <li class="toc-item"><a href="/index.html#depth-level-3">Depth Level 3</a>        <ul class="section-level-3">
             <li class="toc-item"><a href="/index.html#depth-level-4">Depth Level 4</a></li>
-
-                    </ul></li>
 
                     </ul></li>
 
@@ -52,5 +42,3 @@
     </div>
 
 <!-- content end -->
-</body>
-</html>
