+++ tests/Integration/tests/toctree/toctree-glob-level-1/expected/index.html
@@ -13,7 +13,7 @@
             <div class="toc">
     <ul class="menu-level">
     <li class="toc-item"><a href="/page1.html#page-1">Page 1</a></li>
-        <li class="toc-item"><a href="/page2.html#page-1">Page 2</a></li>
+        <li class="toc-item"><a href="/page2.html#page-2">Page 2</a></li>
 
     </ul>
 </div>
